Add a pbs_tracejob scheduler type
pbs_tracejob is similar to the standard pbs scheduler, but uses tracejob
rather than qstat to query for completed jobs.

This is useful for machines that don't support qstat -H. However, the parsing of
tracejob output is likely to be more fragile than parsing qstat -H, so this should be
used with caution.

This is needed for izumi.