cheyenne-nvhpc: Use netcdf/4.8.1 instead of netcdf-mpi/4.8.1
I'm not sure which one is "correct" to use, but I initially had problems
trying to use netcdf-mpi/4.8.1 in the test scripts. I have since found
that I think I could get around those problems by adding an unloadmodule
line and unloading the default netcdf module before the module loads,
but at this point I have done the testing with netcdf/4.8.1, so I'm
going to stick with that for now.

My vague sense is that the netcdf-mpi version supports some additional
things, like parallel access to netcdf files, but I guess you can use
the serial one in mpi applications, at least for most purposes. My vague
sense is that the non-mpi one is more similar to what's in the earlier
netcdf modules that are currently used for other compilers in the
cheyenne nightly testing.

I compared test results with these two mpi modules. For optimized: with
both, all pass except 1 unit test. For debug: results are similar but
not identical: there are slight differences in what passes and what
fails.