Incorporate the creation of the conda env into the test scripts
This way we don't need to explicitly create the conda environment from
the run script. That was problematic because it led to more changes that
were needed to enable ESMPy testing, and required synchronization /
coordination between the run script and config file in terms of the
specific environment(s) being used on a given machine.